fix: cleanup and move update spans function
