+++ clients/web/src/lib/span/update/mutate-span-on-recorded.ts
@@ -1,0 +1,9 @@
+import type { Span } from "~/lib/connection/monitor";
+import type { SpanEvent_Recorded } from "~/lib/proto/spans";
+
+export function mutateSpanOnRecorded(
+  span: Span,
+  recordedEvent: SpanEvent_Recorded,
+) {
+  span.fields = [...span.fields, ...recordedEvent.fields];
+}
